Access control on Notes

+++ backend/src/main/java/eu/projnull/memopad/services/NoteService.java
@@ -1,7 +1,10 @@
 package eu.projnull.memopad.services;
 
 import java.util.List;
+import java.util.Optional;
 
+import eu.projnull.memopad.services.exceptions.NoteDoesntBelongToUser;
+import eu.projnull.memopad.services.exceptions.NoteNotFound;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 
@@ -24,22 +27,30 @@ public class NoteService {
         return noteRepository.save(note);
     }
 
-    public void deleteNote(long ownerId, Note note) {
+    public void deleteNote(long ownerId, Note note) throws NoteDoesntBelongToUser {
+        if (note.getFolder().getOwnerId() != ownerId) throw new NoteDoesntBelongToUser("This note doesn't belong to you!");
         noteRepository.delete(note);
     }
 
-    public Note renameNote(long ownerId, Note note, String newName) {
+    public Note renameNote(long ownerId, Note note, String newName) throws NoteDoesntBelongToUser {
+        if (note.getFolder().getOwnerId() != ownerId) throw new NoteDoesntBelongToUser("This note doesn't belong to you!");
         note.setName(newName);
         return noteRepository.save(note);
     }
 
-    public void moveNote(long ownerId, Note note, Folder newFolder) {
+    public void moveNote(long ownerId, Note note, Folder newFolder) throws NoteDoesntBelongToUser {
+        if (note.getFolder().getOwnerId() != ownerId) throw new NoteDoesntBelongToUser("This note doesn't belong to you!");
         note.setFolder(newFolder);
         noteRepository.save(note);
     }
 
-    public @Nullable Note getNote(long ownerId, long noteId) {
-        return noteRepository.findById(noteId).orElse(null);
+    public @Nullable Note getNote(long ownerId, long noteId) throws NoteDoesntBelongToUser {
+        Optional<Note> noteOptional = noteRepository.findById(noteId);
+        if (noteOptional.isEmpty()) {
+            throw new NoteNotFound("Note with ID " +  noteId + " not found!");
+        }
+        if (noteOptional.get().getFolder().getOwnerId() != ownerId) throw new NoteDoesntBelongToUser("This note doesn't belong to you!");
+        return noteOptional.get();
     }
 
     public Note updateNoteContent(long ownerId, Note note, String content) {
